enterpriseman/ExpertEnterprise/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==7.1.2
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.62.0
h11==0.14.0
idna==3.4
joblib==1.2.0
numpy==1.24.1
pandas==1.5.3
pydantic==1.10.4
PyJWT==1.7.1
python-dateutil==2.8.2
python-decouple==3.3
pytz==2022.7.1
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
starlette==0.13.6
threadpoolctl==3.1.0
typing-extensions==4.4.0
urllib3==1.26.14
uvicorn==0.12.3
